fixed verification issue & added some temp styles
