+++ lib/src/constants.dart
@@ -1,0 +1,3 @@
+import 'package:flutter/material.dart';
+
+const kPrimaryColor = Color(0xff3D4FF4);
