+++ pubspec.yaml
@@ -30,6 +30,7 @@ dependencies:
   cupertino_icons: ^0.1.3
   firebase_auth: ^0.18.0+1
   firebase_core : ^0.5.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -54,19 +55,17 @@ flutter:
   # https://flutter.dev/assets-and-images/#resolution-aware.
 
   # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
+  # https://mflutter.dev/assets-and-images/#from-packages
 
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
   # "family" key with the font family name, and a "fonts" key with a
   # list giving the asset and other descriptors for the font. For
   # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
+  fonts:
+    - family: mont
+      fonts:
+        - asset: assets/fonts/Montserrat-Regular.ttf
   #   - family: Trajan Pro
   #     fonts:
   #       - asset: fonts/TrajanPro.ttf
